适配最新weak转发技术，全面重新整理VC-LTL。

+++ src/msvcrt_winxp.cpp
@@ -633,7 +633,7 @@ EXTERN_C errno_t __cdecl _chsize_s_downlevel(int const fh, __int64 const size)
 {
 	_VALIDATE_CLEAR_OSSERR_RETURN_ERRCODE((size >= 0 && size<=INT_MAX), EINVAL);
 
-	return _chsize(fh, size) ? errno : 0;
+	return _chsize(fh, (long)size) ? errno : 0;
 }
 
 _LCRT_DEFINE_IAT_SYMBOL(_chsize_s_downlevel);
